requirements.txt
streamlit~=1.42.1
python-dotenv
ollama
anthropic
openai# LLM APIs
openai>=1.0.0
anthropic>=0.5.0
ollama>=0.1.0

# Vector Database
weaviate-client>=3.15.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
numpy>=1.22.0
pydantic>=1.9.0

# Web/API
fastapi>=0.95.0
uvicorn>=0.17.0
websockets>=10.3
